Avoid circular dependencies
This commit deletes unneeded imports and fixes the circular dependency
issues.
